feat: extend user login response schema for client request (#19)
* refactor: change Object naming for good conventions.

* test: add more MockData templates

* feat: extend user login response schema for client request

* feat: change details conventions

* fix: add missing changes

* feat: change for request

* refactor: add missing files
