+++ src/main/java/com/modular/restfulserver/auth/application/AuthManageable.java
@@ -1,15 +1,15 @@
 package com.modular.restfulserver.auth.application;
 
-import com.modular.restfulserver.auth.dto.TokenInfo;
 import com.modular.restfulserver.auth.dto.UserLoginRequest;
+import com.modular.restfulserver.auth.dto.UserLoginResponse;
 import com.modular.restfulserver.auth.dto.UserSignupRequest;
 
 import java.util.Map;
 
 public interface AuthManageable {
-  TokenInfo saveUser(UserSignupRequest userSignupRequest);
+  UserLoginResponse saveUser(UserSignupRequest userSignupRequest);
 
-  TokenInfo loginUser(UserLoginRequest userLoginRequest);
+  UserLoginResponse loginUser(UserLoginRequest userLoginRequest);
 
   Map<String, String> refresh(String refreshToken);
