+++ src/main/java/com/modular/restfulserver/user/swagger/UserManagementSchemas.java
@@ -1,6 +1,6 @@
 package com.modular.restfulserver.user.swagger;
 
-import com.modular.restfulserver.user.dto.UserIntegrationInfo;
+import com.modular.restfulserver.user.dto.UserUnitedInfo;
 import com.modular.restfulserver.user.dto.UserInfo;
 import lombok.Getter;
 
@@ -8,7 +8,7 @@ public class UserManagementSchemas {
 
   @Getter
   static class UserInfoSchema {
-    UserIntegrationInfo data;
+    UserUnitedInfo data;
   }
 
   @Getter
