+++ src/test/java/com/modular/restfulserver/user/dto/UserDetailTest.java
@@ -2,18 +2,18 @@ package com.modular.restfulserver.user.dto;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import util.MockData;
+import util.MockProvider;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class UserInfoTest {
+class UserDetailTest {
 
   @Test
-  @DisplayName("[Builder Argument Test] 빌더 패턴으로 생성 시 필드를 제외할 때 예외 발생")
+  @DisplayName("빌더 패턴으로 생성 시 필드를 제외할 때 예외 발생")
   void noAllPostCount() {
     assertThrows(IllegalArgumentException.class, () -> {
-      UserIntegrationInfo.builder()
-        .addUserInfo(MockData.mockUserInfoForClientDto)
+      UserUnitedInfo.builder()
+        .addUserInfo(MockProvider.mockUserInfoForClientDto)
         .addAllGivenLikeCount(0L)
         .addAllFollowingCount(0L)
         .addAllFollowerCount(0L)
